+++ crates/dcc-mcp-http-server/src/lib.rs
@@ -32,10 +32,11 @@ pub mod tool_list_legacy;
 
 // rmcp integration modules (issue #985).
 pub mod rmcp_adapter;
+pub mod rmcp_handler;
 pub mod rmcp_providers;
 pub mod rmcp_registry_context;
+pub mod rmcp_tool_call_async;
 pub mod rmcp_tool_call_dispatch;
-pub mod rmcp_handler;
 
 pub use dynamic_tools::{
     DYNAMIC_TOOL_PREFIX, DynamicToolEntry, DynamicToolError, SessionDynamicTools, ToolSpec,
