+++ crates/dcc-mcp-http-server/src/mcp_tool_list_builder.rs
@@ -4,9 +4,7 @@
 use std::collections::{BTreeMap, HashSet};
 
 use dcc_mcp_gateway::namespace::{BareNameInput, resolve_bare_names};
-use dcc_mcp_jsonrpc::{
-    decode_cursor, encode_cursor, McpTool, TOOLS_LIST_PAGE_SIZE,
-};
+use dcc_mcp_jsonrpc::{McpTool, TOOLS_LIST_PAGE_SIZE, decode_cursor, encode_cursor};
 
 use crate::handlers::build_core_tools;
 use crate::server_state::ServerState;
